refactor: releasesOverview into utils and smaller components for calednar filter
