+++ packages/sanity/src/core/releases/tool/overview/useTimezoneAdjustedDateTimeRange.ts
@@ -1,0 +1,13 @@
+import {endOfDay, startOfDay} from 'date-fns'
+import {useCallback} from 'react'
+
+import useTimeZone from '../../../scheduledPublishing/hooks/useTimeZone'
+
+export const useTimezoneAdjustedDateTimeRange = () => {
+  const {zoneDateToUtc} = useTimeZone()
+
+  return useCallback(
+    (date: Date) => [startOfDay(date), endOfDay(date)].map(zoneDateToUtc),
+    [zoneDateToUtc],
+  )
+}
